refactor: updated the auth function to be reusable
