+++ front/myapp-front/src/components/authForm/AuthFormSignup.tsx
@@ -15,7 +15,7 @@ import { UserContext } from "../../context/user/UserContext";
 import useAxios from "../../hooks/useAxios/useAxios";
 import { PopupContext } from "../../context/popup/PopupContext";
 import { UserAuthRequest } from "../../shared/auth/interfaces";
-import { UserResponse } from "../../shared/user/interfaces";
+import { handleUserAuth } from "../../api/auth/api";
 
 function AuthFormSignup({ setIsLogin }: AuthFormProps) {
   const { dispatch: userDispatch } = useContext(UserContext);
@@ -36,28 +36,6 @@ function AuthFormSignup({ setIsLogin }: AuthFormProps) {
         },
       };
     });
-  }
-
-  async function handleSubmit() {
-    const response = await handleAxiosCall<UserAuthRequest, UserResponse>({
-      method: "post",
-      url: "http://localhost:3000/users",
-      data: data,
-      needAuth: false,
-    });
-
-    if (response) {
-      const token = response.headers.authorization.split(" ")[1];
-      // To set the Cookie
-      setAuthToken(token);
-      // To set the token in the context
-      userDispatch({ type: "SET_USER_AUTH", payload: token });
-      // To set the user info in the context
-      userDispatch({ type: "SET_USER", payload: response.data.userObject });
-
-      // Add notification
-      setPopup({ message: response.data.message, type: "success" });
-    }
   }
 
   return (
@@ -126,7 +104,16 @@ function AuthFormSignup({ setIsLogin }: AuthFormProps) {
       />
       <Button
         variant="contained"
-        onClick={handleSubmit}
+        onClick={() =>
+          handleUserAuth({
+            handleAxiosCall,
+            data,
+            url: "http://localhost:3000/users",
+            setAuthToken,
+            userDispatch,
+            setPopup,
+          })
+        }
         sx={{
           color: "#515151",
           fontWeight: "bold",
